requirements.txt
flask
torch
torchvision
timm
numpy
Pillow
git+https://github.com/ZiyaoLi/fast-kan.git